+++ js/firebase/auth.js
@@ -1,27 +1,3 @@
-import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
-import { app } from "./config";
-
-const auth = getAuth()
-
-const loginForm = document.querySelector("#login-form")
-
-loginForm.addEventListener("submit", (e) => {
-  e.preventDefault()
-  const email = document.querySelector("#email").value
-  const pwd = document.querySelector("#pwd").value
-
-  signInWithEmailAndPassword(auth, email, pwd)
-    .then((credential) => {
-      // get info about user
-      const user = credential.user;
-      console.log(user)   // test to see if sign in works
-    })
-    .catch((error) => {
-      // see error
-      console.log(error.code)
-      console.log(error.message)
-    });
-})
 
 
 //
