+++ test.html
@@ -24,7 +24,7 @@
     <input type="submit" name="login" value="Login" id="login-btn">
   </form>
 
-  <script type="module" src="./js/firebase/config.js"></script>
-  <script type="module" src="./js/firebase/auth.js"></script>
+  <script src="bundle.js"></script>
+
 </body>
 </html>
